% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_tibble.mcmc_data}
\alias{as_tibble.mcmc_data}
\title{As Tibble}
\usage{
\method{as_tibble}{mcmc_data}(x, samples = FALSE, ...)
}
\arguments{
\item{x}{An mcmc_data object to coerce to a tibble.}

\item{samples}{A flag specifying whether to include samples}

\item{...}{Unused, for extensibility.}
}
\description{
As Tibble
}
