% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-rows.R
\name{bind_rows2}
\alias{bind_rows2}
\title{Bind Rows}
\usage{
bind_rows2(x, y)
}
\arguments{
\item{x}{A mcmc_data}

\item{y}{A second mcmc_data}
}
\description{
Bind Rows
}
\examples{
bind_rows2(mcmc_data_example, mcmc_data_example)
}
