% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mcmc_data}
\alias{check_mcmc_data}
\title{Check mcmc data}
\usage{
check_mcmc_data(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object x.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks an object is a \code{mcmc_data}.
}
\examples{
check_mcmc_data(mcmcdata::mcmc_data_example)
}
