% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mcmc_data}
\alias{coef.mcmc_data}
\title{Coefficients}
\usage{
\method{coef}{mcmc_data}(object, conf_level = 0.95,
  estimate = stats::median, ...)
}
\description{
Coefficients
}
\examples{
coef(mcmc_data_example)
}
