% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_by.mcmc_data}
\alias{group_by.mcmc_data}
\title{Group mcmc data}
\usage{
\method{group_by}{mcmc_data}(.data, ..., .add = FALSE)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{In \code{group_by()}, variables or computations to group by.
In \code{ungroup()}, variables to remove from the grouping.}

\item{.add}{When \code{FALSE}, the default, \code{group_by()} will
override existing groups. To add to the existing groups, use
\code{.add = TRUE}.

This argument was previously called \code{add}, but that prevented
creating a new grouping variable called \code{add}, and conflicts with
our naming conventions.}
}
\description{
Group mcmc data
}
\examples{
group_by(mcmc_data_example, homeworld, species)
}
