% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmc_data}
\alias{is.mcmc_data}
\title{Is an MCMC data object}
\usage{
is.mcmc_data(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mcmc_data'
}
