% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-data.R
\name{mcmc_data}
\alias{mcmc_data}
\title{MCMC Data}
\usage{
mcmc_data(mcmc, data)
}
\arguments{
\item{mcmc}{An mcmc object with one parameter}

\item{data}{A data frame}
}
\value{
An object of class \code{mcmc_data}
}
\description{
Combines an mcmc object with a data frame. 
The mcmc object must be for a single parameter of the same length as the number of rows in the data frame.
The resultant object can be filtered or summarized or the coef calculated.
}
