% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_data_example}
\alias{mcmc_data_example}
\title{Example mcmc data Object}
\format{An object of class \code{mcmc_data} of length 2.}
\usage{
mcmc_data_example
}
\description{
An example mcmc_data object.
}
\keyword{datasets}
