% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-derive.R
\name{mcmc_derive.mcmc_data}
\alias{mcmc_derive.mcmc_data}
\title{mcmc_derive mcmc_data}
\usage{
\method{mcmc_derive}{mcmc_data}(
  object,
  expr = "derived <- current",
  values = list(),
  parameter = "current",
  monitor = "^derived$",
  parallel = TRUE,
  ...
)
}
\arguments{
\item{object}{An mcmc_data object}

\item{expr}{A string of the R expression.}

\item{values}{A named list of values for objects not in object$data}

\item{parameter}{A string of the name of the current MCMC samples in expr.}

\item{monitor}{A string of the}

\item{parallel}{A flag specifying whether to run in parallel.}

\item{...}{An unused.}
}
\description{
mcmc_derive mcmc_data
}
