% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-derive-data.R
\name{mcmc_derive_data.mcmc_data}
\alias{mcmc_derive_data.mcmc_data}
\title{mcmc_derive mcmc_data}
\usage{
\method{mcmc_derive_data}{mcmc_data}(
  object,
  expr = "new_par <- par",
  values = list(),
  parameter = "par",
  monitor = "new_par",
  parallel = FALSE,
  silent = getOption("mcmcderive.silent", FALSE),
  ...
)
}
\arguments{
\item{object}{An mcmc_data object}

\item{expr}{A string of the R expression.}

\item{values}{A named list of values for objects not in object$data}

\item{parameter}{A string of the name of the current MCMC samples in expr.}

\item{monitor}{A string of the name of the derived parameter to return.}

\item{parallel}{A flag specifying whether to generate the derived parameters
for each chain in parallel.}

\item{silent}{A flag specifying whether to suppress messages and warnings.}

\item{...}{Unused.}
}
\value{
An mcmc_data object.
}
\description{
Gets mcmc_data object of a derived parameters.
To return an mcmcr object of one or more derived parameters use mcmc_derive().
By default the current parameter is call par and the derived one is called new_par.
}
