% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-derive-data.R
\name{mcmc_derive_data}
\alias{mcmc_derive_data}
\title{Derive MCMC Data}
\usage{
mcmc_derive_data(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Not used.}
}
\description{
Get mcmc_data object.
}
