% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/fill-na.R, R/filter.R,
%   R/group-by.R, R/join.R, R/mcmc-derive.R, R/mutate.R, R/rename.R,
%   R/replace-na.R, R/select.R, R/slice.R, R/summarise.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{fill_na}
\alias{filter}
\alias{group_by}
\alias{ungroup}
\alias{anti_join}
\alias{semi_join}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{mcmc_derive}
\alias{mutate}
\alias{rename}
\alias{replace_na}
\alias{select}
\alias{slice}
\alias{summarise}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{extras}{\code{\link[extras]{fill_na}}}

  \item{mcmcderive}{\code{\link[mcmcderive]{mcmc_derive}}}

  \item{tidyr}{\code{\link[tidyr]{replace_na}}}
}}

