% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.mcmc_data}
\alias{select.mcmc_data}
\title{Select mcmcr data}
\usage{
\method{select}{mcmc_data}(.data, ...)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\description{
Select mcmcr data
}
\examples{
select(mcmc_data_example, species, homeworld)
}
