% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{ungroup.mcmc_data}
\alias{ungroup.mcmc_data}
\title{Ungroup mcmc data}
\usage{
\method{ungroup}{mcmc_data}(x, ...)
}
\arguments{
\item{x}{An mcmc_data object}

\item{...}{Variables to group by. All tbls accept variable names.
Some tbls will accept functions of variables. Duplicated groups
will be silently dropped.}
}
\description{
Ungroup mcmc data
}
\examples{
ungroup(mcmc_data_example)
}
