% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange.mcmc_data}
\alias{arrange.mcmc_data}
\title{Arrange mcmcr data}
\usage{
\method{arrange}{mcmc_data}(.data, ...)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Variables, or functions or
variables. Use \code{\link[dplyr:desc]{desc()}} to sort a variable in descending order.}
}
\description{
Arrange mcmcr data
}
\examples{
arrange(mcmc_data_example, species, homeworld)
}
