% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.mcmc_data}
\alias{filter.mcmc_data}
\title{Slice mcmcr data}
\usage{
\method{filter}{mcmc_data}(.data, ...)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{For \code{slice()}: <\code{\link[rlang:args_data_masking]{data-masking}}>
Integer row values.

Provide either positive values to keep, or negative values to drop.
The values provided must be either all positive or all negative.
Indices beyond the number of rows in the input are silently ignored.

For \verb{slice_*()}, these arguments are passed on to methods.}
}
\description{
Slice mcmcr data
}
\examples{
filter(mcmc_data_example, species == "Droid")
}
