% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-na.R
\name{replace_na.mcmc_data}
\alias{replace_na.mcmc_data}
\title{Replace Missing Values mcmc data}
\usage{
\method{replace_na}{mcmc_data}(data, replace, ...)
}
\arguments{
\item{data}{A data frame or vector.}

\item{replace}{If \code{data} is a data frame, \code{replace} takes a list of values,
with one value for each column that has \code{NA} values to be replaced.

If \code{data} is a vector, \code{replace} takes a single value. This single value
replaces all of the \code{NA} values in the vector.}

\item{...}{Additional arguments for methods. Currently unused.}
}
\description{
Replace Missing Values mcmc data
}
