% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.mcmc_data}
\alias{select.mcmc_data}
\title{Select mcmcr data}
\usage{
\method{select}{mcmc_data}(.data, ...)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{One or more unquoted expressions separated by commas.
You can treat variable names like they are positions.

Positive values select variables; negative values to drop variables.
If the first expression is negative, \code{select()} will automatically
start with all variables.

Use named arguments to rename selected variables.

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in a context where column names
represent column positions. They support
\link[rlang:quasiquotation]{unquoting} and splicing. See
\code{vignette("programming")} for an introduction to these concepts.}
}
\description{
Select mcmcr data
}
\examples{
select(mcmc_data_example, species, homeworld)
}
