% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice.mcmc_data}
\alias{slice.mcmc_data}
\title{Slice mcmcr data}
\usage{
\method{slice}{mcmc_data}(.data, ...)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{Integer row values.  Provide either positive values to keep,
or negative values to drop. The values provided must be either all
positive or all negative.  Indices beyond the number of rows in the
input are silently ignored.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\description{
Slice mcmcr data
}
\examples{
slice(mcmc_data_example, 1L)
}
