% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise.mcmc_data}
\alias{summarise.mcmc_data}
\title{Summarise mcmc data}
\usage{
\method{summarise}{mcmc_data}(.data, ..., .fun = sum)
}
\arguments{
\item{.data}{An mcmc_data object}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs of
summary functions. The name will be the name of the variable in the result.

The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A data frame, to add multiple columns from a single expression.
}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Returning values with size 0 or >1 was
deprecated as of 1.1.0. Please use \code{\link[dplyr:reframe]{reframe()}} for this instead.}

\item{.fun}{The function to use to summarise the MCMC samples.}
}
\description{
Summarise mcmc data
}
